[ORC] Fix weak hidden symbols failure on PPC with runtimedyld
Fix "JIT session error: Symbols not found: [ DW.ref.__gxx_personality_v0 ] error" which happens when trying to use exceptions on ppc linux. To do this, it adds AutoClaimWeakHiddenSymbols option in RTDyldObjectLinkingLayer. This option claims responsibility for every weak hidden symbol if it didn't get added to responsibility set before in IRLayer/ObjectLayer step. In ppc linux, DW.ref symbols is emitted as weak hidden symbols in the later stage of MC pipeline. This means when using IRLayer (i.e. LLJIT), IRLayer will not claim responsibility for such symbols and RuntimeDyld will skip defining this symbol even though it couldn't resolve corresponding external symbol. Existing AutoClaimResponsibility option will not work too because the symbol is weak hidden and it will not be added to Resolved symbol set. Hence, the motivation for this option.

Differential Revision: https://reviews.llvm.org/D129175